Rename landmark to landmark_score
