+++ lib/compare_all.js
@@ -6,7 +6,7 @@ var placeSignificant = require('../comparators/place_significant');
 var userChangesets = require('../comparators/user_changesets');
 var countTag = require('../comparators/count_tag');
 var deleteCreate = require('../comparators/delete_create');
-var landmark = require('../comparators/landmark.js');
+var landmarkScore = require('../comparators/landmark_score.js');
 var cityDeleted = require('../comparators/city_deleted');
 var compareProperties = require('../comparators/compare_properties');
 var compareGeometries = require('../comparators/compare_geometries');
@@ -31,7 +31,7 @@ var compareAll = function(newVersion, oldVersion, callback) {
     placeSignificant,
     userChangesets,
     deleteCreate,
-    landmark,
+    landmarkScore,
     cityDeleted,
     compareProperties,
     compareGeometries,
